v2.20
RELEASE NOTES
=============

v2.20 - 29 8/27/2004

NEW FEATURES
------------

FEATURE: Polling
On startup by user action the app polls for new videos. This is handy
in case the phone is in a coverage area where it can do http request
but is not able to receive notifications, or missed a notification that
was not resent when it was out of coverage.

FEATURE: Icon installed on the home screen the first time you register.

BUG FIXES
---------

PUNTED BUG FIX: App flashing when intents that should run in the
background come in. The work around for android bug:
https://code.google.com/p/android/issues/detail?id=4394 would require a
very invasive change to the architecture. Cant seem to find another fix
so punting on this for the time being.