NineViewGroup more cleanup. NineViewGroup.GestureListener cleanup.
NineView group cleanUp
- Test no measure
- Fix position ordering.
- Add private methods postionWithIndex() indexWithPostion() they are never used outside of nineViewGroup
- Nineview group lays out and provides references to nine FrameLayouts

NineViewGroup.GestureListener
- Did some clean up to eliminate calls to adapter.
- It is ok for now.
- But this code needs to be refactored as a separate LongPressGestureRecognizer class as I had it
before. The constructors for this class should pass it a main background view on which to process touch events and a set of target views.
It should fire events and return a target view with which the event is associated exactly like I had LongpressTouchHandler in the original code.